+++ tests/core/test_proxy.py
@@ -1,5 +1,6 @@
 from unittest.mock import MagicMock, patch
 from contextlib import ExitStack
+import logging
 
 import pytest
 from PySide6.QtCore import QMutex
@@ -101,9 +102,39 @@ def test_proxyExists(proxy):
     proxy.proxy_path = "/proxy/path/proxy.png"
     assert proxy.proxyExists()
 
-def test_cleanup(proxy):
-    proxy.proxy_path = "/proxy/path/proxy.png"
+def test_cleanUp_no_proxy(proxy):
+    assert proxy.proxy_path is None
     with patch("core.proxy.os.remove") as mock_remove:
-        proxy.cleanup()
-        mock_remove.assert_called_once_with("/proxy/path/proxy.png")
+        proxy.cleanUp()
+        mock_remove.assert_not_called()
+
+def test_cleanUp_happy_path(proxy):
+    proxy_file = "/proxy/path/proxy.png"
+    proxy.proxy_path = proxy_file
+    with patch("core.proxy.os.remove") as mock_remove:
+        proxy.cleanUp(raising=True)
+        mock_remove.assert_called_once_with(proxy_file)
         assert proxy.proxy_path is None
+
+def test_cleanUp_sad_path_raising(proxy):
+    proxy_file = "/proxy/path/proxy.png"
+    proxy.proxy_path = proxy_file
+    with (
+        patch("core.proxy.os.remove", side_effect=OSError) as mock_remove,
+        pytest.raises(FileException),
+    ):
+        proxy.cleanUp(raising=True)
+        mock_remove.assert_called_once_with(proxy_file)
+        assert proxy.proxy_path is None
+
+def test_cleanUp_sad_path_not_raising(proxy, caplog):
+    proxy_file = "/proxy/path/proxy.png"
+    proxy.proxy_path = proxy_file
+    with (
+        patch("core.proxy.os.remove", side_effect=OSError) as mock_remove,
+        caplog.at_level(logging.ERROR)
+    ):
+        proxy.cleanUp(raising=False)
+        mock_remove.assert_called_once_with(proxy_file)
+        assert proxy.proxy_path is None
+        assert "Failed to clean up proxy" in caplog.text